backend/services/location-service/requirements.txt
# Location Service Dependencies
# ProActive People Recruitment Platform - UK Postcode & Geography API

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database - PostgreSQL
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Environment & Configuration
python-dotenv==1.0.0

# File Upload
python-multipart==0.0.6

# Utilities
regex==2023.12.25

# Development & Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0

# Logging & Monitoring
python-json-logger==2.0.7
